requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
Pillow==10.2.0
psycopg2-binary==2.9.9
